feat: Add research template
For Infrastructure work at the very least, we often encounter the need
to work out how we want to go forward with something, but working it out
is quite a task in and of itself. So, like most things, we want to be
able to track the research work along with our other tasks.